[dart] Remove workaround now that type promotion accounts for local boolean variables.
This change removes a workaround that was introduced prior to landing
Dart language feature
https://github.com/dart-lang/language/issues/1274, which allows type
promotion in null safe code to account for local boolean variables.
The workaround ensured that the code would analyze the same regardless
of whether the feature was enabled, allowing for a smoother
transition.  Now that the feature has fully landed, the workaround
isn't needed anymore.

Change-Id: I5338bd5e25c977452881239e5d5b843c80a59b31
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/474237
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Chase Latta <chaselatta@google.com>