[storage] Change io2's Unlink method.
The unlink method needs some options so that clients can assert the
target is a directory.

We also change the method so that it specifies a name rather than a
path.  All our current implementations don't support path traversal
server-side, so this shouldn't be an issue.  If we need to support
traversal later (perhaps for performance reasons), we can change it back
and since it would be implemented within VFS, it shouldn't be too
problematic to change it.

Bug: 74923
Bug: 74911
Bug: 75137
Change-Id: I33ecbac14289367ad148313805f2aa40e74ed3cd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/518063
Commit-Queue: Chris Suter <csuter@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Reviewed-by: Scott Graham <scottmg@google.com>
API-Review: Adam Barth <abarth@google.com>