[fidl][dart] Receive-side handle rights checking
This CL adds receive-side handle rights checking in Dart. This means
that handle rights and type are checked on the receiving side of the
channel.

The tests in this CL just test end-to-end behavior and do no isolate
checks that the behavior is happening on the receiving side (that will
happen later). In addition, they only check that the _checkHandleRights
method is called by testing that wrong handle types are caught because
detailed tests of type and rights variations will be covered by GIDL
decode tests.

Some upcoming changes:
- zx_handle_replace is not currently supported in Dart, but it will be
supported after the next Flutter roll. Because of this, rights reduction
cannot be supported at this time.
- GIDL support will be used to validate the decoding logic. However,
this involves significant changes and will be done in a separate CL.
- Send-side handle rights checking needs to be completed as well.

Bug: 41920
Change-Id: I54ca20e7f8d8b1883f9cb63c9df4cfabdb1e8b67
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/534541
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>