[fidl][dart] Reduce handle rights when there are extra rights
When more than the required rights are passed to a FIDL decoder, the
decoder is expected to reduce the rights. This CL implements this logic
in Dart.

The test is written without GIDL because sufficient capability to
implement this in GIDL doesn't exist yet in dart (zx_object_get_info).

Change-Id: I94c7e99736bec4d2c32164f48e1ab24a9792fa88
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/542263
Reviewed-by: Mitchell Kember <mkember@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>