[fidl][dart] Wire format v2 encoding for dart
Previously, support was added for decoding FIDL wire format v2 in dart.
This CL adds the other direction -- encode is now supported.

This functionality is not turned on by default and is gated by a wire
format type that is passed into the encoder.

GIDL encode tests are run against this implementation.

Change-Id: Ib4cc4de426b4869db6513aa697196de5119e8b2f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/568721
Reviewed-by: Ian McKellar <ianloic@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>