[flutter] Fix runner crash when switching between views
This change fixes the crash observed in the flutter runner when the user
switches views on the Workstation shell. This is because the view was
destroyed in FuchsiaViewController.dispose, when the PlatformView
was taken Offstage. This is flutter behavior when a PlatformView is
moved offstage when switching. Calling view destroy here is wrong because the
underlying scenic view and the owning component/app is still alive and
the user is expected to switch back to the app.

There is no use case to call destroy view method on the
FuchsiaViewsService. Rather, the runner notifies the client code that
the view was destroyed in response to the underlying scenic session
closing. The method should be removed to reduce developer confusion.

Fixed: 83841
Test: fx test --host fuchsia_scenic_flutter_unittests
Change-Id: I73e31d3cd8465731c8bbbf0dec965973a3db7746
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/579147
Reviewed-by: Yeonhee Lee <yhlee@google.com>
Reviewed-by: David Worsham <dworsham@google.com>
Commit-Queue: Sanjay Chouksey <sanjayc@google.com>