[fidl][dart] Validate padding, disallow 0 ordinal
In the generated struct decode function, we now generate check padding
statements for all byte intervals which must be zero. In addition, we
check the padding which must be added at the end of inline structures
which are not FIDL aligned, i.e. whose size is not exactly a multiple of
eight.

We also disallow the zero ordinal in unions.

Bug: 7805
Change-Id: I735ffbf9a076389b2df581a7b5525b9b8b40a84c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/607824
Commit-Queue: Pascal-Louis Perez <pascallouis@google.com>
Fuchsia-Auto-Submit: Pascal-Louis Perez <pascallouis@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>