[build][testing] Add more missing package deps
This CL adds a large number of the remaining package dependencies in
existing tests. It was generated by running the script in
https://fxrev.dev/606442 iteratively until no further missing deps were
found. The list was then curated slightly to account for some out of
order logs in the syslog, and updated with deps that did not match the
usual resolution failure signature.

Note that this CL still does not bring us to 100% fully specified
package deps. There are two remaining root causes that have yet to be
solved:
1. Tests utilizing prebuilt packages shipped by chrome do not correctly
specify deps, as that will require some refactoring of the visibility
rules in the associated package templates.
2. Host tests do not correctly specify package deps as those templates
have not been updated to handle this field.
Both of these root causes will be addressed in upcoming CLs, as this one
is already quite large.

Bug: b/195343539
Change-Id: I1e12d67a0e6b4f2218387dc1a02d16ee8e708745
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/607345
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>