[fuchsia.io] Subsume fuchsia.io2
Separate protocols in fuchsia.io into 1 and 2 trees; e.g. Node now
composes Node1 and Node2, Directory composes Directory1 and Directory2,
Directory1 composes Node1, and Directory2 composes Node2. This allows
the migration to be tracked more obviously; the migration of a protocol
is complete when its "1" variant is empty.

This uncovered some oddities:
- Because AdvisoryLocking was composed into io from io2, its ordinal was
  computed from its natural selector. This means it now has a synthetic
  "io2" selector in io.
- Directory.AddInotifySelector has an "io1" selector, though it is part
  of io2.
- File.GetMemRange had become File.GetBackingMemory in fuchsia.io.
- NodeProtocols::{Debuglog,PosixSocket} were replaced with
  NodeProtocols::{Stream,Datagram,Raw}Socket in fuchsia.io.
- fuchsia.io2::DeviceSignal values diverged from
  fuchsia.io.DEVICE_SIGNAL_* values; discovered because Go bindings
  generate the same name for both patterns, causing collisions.

The primary downside of this approach is that all "2" types now include
"2" in their names, which affects generated code. A similar choice was
already made with overlapping method names (e.g. Close2), so the
implication (a global rename of the world at the end of this migration)
is judged to be acceptable.

fuchsia.io2 is temporarily kept in place as a stub containing only
`Operations` to allow usage in another repository to soft transition.

Bug: 77623
Change-Id: If25aef423d96e1e2e999a6c86100976584576ec9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/642884
Commit-Queue: Tamir Duberstein <tamird@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
API-Review: Adam Barth <abarth@google.com>
Commit-Queue: Charles Celerier <chcl@google.com>