[dart] Enable Dart API checking
The script "detect_api_changes" will now be run on all dart libraries
that are included in the SDK.

If any changes are detected, an error will
appear in CQ instructing the user to update the associated .api file.

Then, with a change to the .api file, an API reviewer will be required
to submit the change.

Change-Id: I586e01d16aff0dac9ef54d6cadf2260911d63ce9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/644269
Reviewed-by: Chase Latta <chaselatta@google.com>
Reviewed-by: Misha Gridnev <gridman@google.com>
API-Review: Chase Latta <chaselatta@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>