[sdk-dart][realm-builder] child to child routing, events, and bug fixes
Connecting to services through ScopedInstance was incorrectly closing
the exposedDir, so it could not be used again. This also affected the
`connectToBinder()` call, if called after connecting to another service
via the `exposedDir`.

A new test that shows a realm with separate client and server components
requires starting the client eagerly. Since the test is not involved,
the realm has to be started, via `connectToBinder()`.

The only way to validate the test succeeds is to catch the component
`stopped` event (which should happen after the client starts, calls
`echoString()` on the server, and gets the server's response. So this
test shows how to connect to the framework `EventSource`, subscribe,
listen for the stopped event, and check for a successful status.

Change-Id: I8a5effa692de5221cd8344594e400b4551c5b8f4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/661895
Reviewed-by: Derek Gonyeo <dgonyeo@google.com>
Reviewed-by: Chase Latta <chaselatta@google.com>
Commit-Queue: Rich Kadel <richkadel@google.com>